U-Net/requirements.txt
IPython[all]
tensorflow==2.8.0
opencv-python==4.7.0.72
pillow==9.5.0
matplotlib==3.7.1
protobuf==3.20.0
